move format_problem to utils
