+++ srl/commands/nextup/add.py
@@ -1,5 +1,5 @@
 from rich.console import Console
-from srl.commands.list_ import format_problem
+from srl.utils import format_problem
 from srl.utils import today
 from srl.storage import (
     load_json,
