+++ srl/commands/nextup/list_.py
@@ -1,6 +1,6 @@
 from rich.console import Console
 from rich.panel import Panel
-from srl.commands.list_ import format_problem
+from srl.utils import format_problem
 from .util import get_next_up_problems
